feat: Show gap stocks

+++ lib/polygon/types.go
@@ -60,17 +60,17 @@ type TickerBar struct {
 
 type TickerOHLC struct {
 	Volume float64 `json:"v"`
-	Open   float32 `json:"o"`
-	High   float32 `json:"h"`
-	Low    float32 `json:"l"`
-	Close  float32 `json:"c"`
+	Open   float64 `json:"o"`
+	High   float64 `json:"h"`
+	Low    float64 `json:"l"`
+	Close  float64 `json:"c"`
 }
 
 type TickerSnapshot struct {
 	Ticker             string     `json:"ticker"`
 	Day                TickerOHLC `json:"day"`
 	PrevDay            TickerOHLC `json:"prevDay"`
-	Change             float32    `json:"todaysChange"`
-	ChangePercent      float32    `json:"todaysChangePerc"`
+	Change             float64    `json:"todaysChange"`
+	ChangePercent      float64    `json:"todaysChangePerc"`
 	UpdatedAtUnixNanos int64      `json:"updated"`
 }
